requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
better-ipc==2.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
chat_exporter==2.7.1
click==8.1.7
colorama==0.4.6
emoji==2.11.1
fastapi==0.110.0
ffmpeg==1.4
frozenlist==1.4.1
function-cooldowns==2.0.1
grapheme==0.6.0
gTTS==2.5.1
h11==0.16.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mysql-connector-python==9.1.0
nextcord @ git+https://github.com/TrapDrap/_nextcord@a00b3c6aa7d4ed0e9c01d238a26fda6c9f10367f
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyreadline3==3.4.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.0
rich==13.7.1
sniffio==1.3.1
typing_extensions==4.10.0
urllib3==2.2.2
uvicorn==0.28.0
websockets==12.0
yarl==1.9.4
